requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
gevent>=24.2.1
gevent-websocket>=0.10.1
redis==4.5.4
Pillow==9.5.0